+++ src/main/java/dev/doctor4t/ratatouille/client/model/armor/CustomBipedArmorModel.java
@@ -1,69 +1,62 @@
 package dev.doctor4t.ratatouille.client.model.armor;
 
-import net.minecraft.client.model.Dilation;
-import net.minecraft.client.model.ModelData;
-import net.minecraft.client.model.ModelPart;
-import net.minecraft.client.render.VertexConsumer;
+import net.minecraft.client.model.*;
 import net.minecraft.client.render.entity.model.BipedEntityModel;
-import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.client.render.entity.model.EntityModelPartNames;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.util.Identifier;
 
 import java.util.function.Consumer;
 
 public class CustomBipedArmorModel<T extends LivingEntity> extends BipedEntityModel<T> {
-    public final ModelPart head;
-    public final ModelPart body;
+    private final CustomArmorModelDefinition modelDefinition;
+
+    public final ModelPart helmet;
     public final ModelPart body_chestplate;
     public final ModelPart body_leggings;
-    public final ModelPart right_arm;
-    public final ModelPart left_arm;
-    public final ModelPart right_leg;
+    public final ModelPart right_arm_chestplate;
+    public final ModelPart left_arm_chestplate;
     public final ModelPart right_leg_leggings;
     public final ModelPart right_leg_boot;
-    public final ModelPart left_leg;
     public final ModelPart left_leg_leggings;
     public final ModelPart left_leg_boot;
-    private final CustomArmorModelDefinition modelDefinition;
 
     public CustomBipedArmorModel(ModelPart root, CustomArmorModelDefinition modelDefinition) {
         super(root);
 
         this.modelDefinition = modelDefinition;
 
-        this.head = root.getChild("head");
+        this.helmet = this.head.getChild("helmet");
 
-        this.body = root.getChild("body");
         this.body_leggings = this.body.getChild("body_leggings");
         this.body_chestplate = this.body.getChild("body_chestplate");
 
-        this.right_arm = root.getChild("right_arm");
+        this.right_arm_chestplate = this.rightArm.getChild("right_arm_chestplate");
 
-        this.left_arm = root.getChild("left_arm");
+        this.left_arm_chestplate = this.leftArm.getChild("left_arm_chestplate");
 
-        this.right_leg = root.getChild("right_leg");
-        this.right_leg_leggings = this.right_leg.getChild("right_leg_leggings");
-        this.right_leg_boot = this.right_leg.getChild("right_leg_boot");
+        this.right_leg_leggings = this.rightLeg.getChild("right_leg_leggings");
+        this.right_leg_boot = this.rightLeg.getChild("right_leg_boot");
 
-        this.left_leg = root.getChild("left_leg");
-        this.left_leg_leggings = this.left_leg.getChild("left_leg_leggings");
-        this.left_leg_boot = this.left_leg.getChild("left_leg_boot");
+        this.left_leg_leggings = this.leftLeg.getChild("left_leg_leggings");
+        this.left_leg_boot = this.leftLeg.getChild("left_leg_boot");
     }
 
     public static ModelData getModelData(Consumer<ModelData> modelDefinition, Dilation dilation) {
-        ModelData modelData = BipedEntityModel.getModelData(dilation, 0.0F);
-        modelDefinition.accept(modelData);
-        return modelData;
-    }
+        ModelData modelData = new ModelData();
 
-    @Override
-    public void render(MatrixStack matrices, VertexConsumer vertices, int light, int overlay, int color) {
-        head.render(matrices, vertices, light, overlay, color);
-        body.render(matrices, vertices, light, overlay, color);
-        right_arm.render(matrices, vertices, light, overlay, color);
-        left_arm.render(matrices, vertices, light, overlay, color);
-        right_leg.render(matrices, vertices, light, overlay, color);
-        left_leg.render(matrices, vertices, light, overlay, color);
+        ModelPartData modelPartData = modelData.getRoot();
+        modelPartData.addChild(EntityModelPartNames.HEAD, ModelPartBuilder.create(), ModelTransform.pivot(0.0F, 0.0F, 0.0F));
+        modelPartData.addChild(EntityModelPartNames.HAT, ModelPartBuilder.create(), ModelTransform.pivot(0.0F, 0.0F, 0.0F));
+        modelPartData.addChild(EntityModelPartNames.BODY, ModelPartBuilder.create(), ModelTransform.pivot(0.0F, 0.0F, 0.0F));
+        modelPartData.addChild(EntityModelPartNames.RIGHT_ARM, ModelPartBuilder.create(), ModelTransform.pivot(0.0F, 0.0F, 0.0F));
+        modelPartData.addChild(EntityModelPartNames.LEFT_ARM, ModelPartBuilder.create(), ModelTransform.pivot(0.0F, 0.0F, 0.0F));
+        modelPartData.addChild(EntityModelPartNames.RIGHT_LEG, ModelPartBuilder.create(), ModelTransform.pivot(0.0F, 0.0F, 0.0F));
+        modelPartData.addChild(EntityModelPartNames.LEFT_LEG, ModelPartBuilder.create(), ModelTransform.pivot(0.0F, 0.0F, 0.0F));
+
+        modelDefinition.accept(modelData);
+
+        return modelData;
     }
 
     public Identifier getTexture() {