+++ src/main/java/dev/doctor4t/ratatouille/client/render/feature/RendersArmInFirstPerson.java
@@ -1,6 +1,5 @@
 package dev.doctor4t.ratatouille.client.render.feature;
 
-import net.minecraft.client.model.ModelPart;
 import net.minecraft.client.network.AbstractClientPlayerEntity;
 import net.minecraft.client.render.entity.model.BipedEntityModel;
 import net.minecraft.entity.LivingEntity;
@@ -12,8 +11,4 @@ public interface RendersArmInFirstPerson<T extends LivingEntity> {
     BipedEntityModel<T> getModel(LivingEntity livingEntity);
 
     Identifier getTexture(LivingEntity livingEntity);
-
-    ModelPart getRightArm(LivingEntity livingEntity);
-
-    ModelPart getLeftArm(LivingEntity livingEntity);
 }